move ceil_log2 into o1_utils::math package
