+++ kimchi/src/tests/poseidon.rs
@@ -5,7 +5,7 @@ use ark_ff::{UniformRand, Zero};
 use ark_poly::{univariate::DensePolynomial, UVPolynomial};
 use array_init::array_init;
 use colored::Colorize;
-use commitment_dlog::commitment::{b_poly_coefficients, ceil_log2, CommitmentCurve};
+use commitment_dlog::commitment::{b_poly_coefficients, CommitmentCurve};
 use groupmap::GroupMap;
 use mina_curves::pasta::{
     fp::Fp,
@@ -29,6 +29,8 @@ use crate::{
 };
 use crate::{prover::ProverProof, prover_index::ProverIndex};
 
+use o1_utils::math;
+
 // aliases
 
 type SpongeParams = PlonkSpongeConstantsKimchi;
@@ -46,7 +48,7 @@ const N_LOWER_BOUND: usize = (POS_ROWS_PER_HASH + 1) * NUM_POS; // Plonk domain 
 
 #[test]
 fn test_poseidon() {
-    let max_size = 1 << ceil_log2(N_LOWER_BOUND);
+    let max_size = 1 << math::ceil_log2(N_LOWER_BOUND);
     println!("max_size = {}", max_size);
     println!("rounds per hash = {}", ROUNDS_PER_HASH);
     println!("rounds per row = {}", ROUNDS_PER_ROW);
@@ -85,7 +87,7 @@ fn test_poseidon() {
 /// creates a proof and verifies it
 fn positive(index: &ProverIndex<Affine>) {
     // constant
-    let max_size = 1 << ceil_log2(N_LOWER_BOUND);
+    let max_size = 1 << math::ceil_log2(N_LOWER_BOUND);
 
     // set up
     let rng = &mut StdRng::from_seed([0u8; 32]);
@@ -136,7 +138,7 @@ fn positive(index: &ProverIndex<Affine>) {
 
         //
         let prev = {
-            let k = ceil_log2(index.srs.g.len());
+            let k = math::ceil_log2(index.srs.g.len());
             let chals: Vec<_> = (0..k).map(|_| Fp::rand(rng)).collect();
             let comm = {
                 let coeffs = b_poly_coefficients(&chals);
