+++ utils/src/lib.rs
@@ -2,6 +2,7 @@ pub mod dense_polynomial;
 pub mod evaluations;
 pub mod field_helpers;
 pub mod hasher;
+pub mod math;
 pub mod serialization;
 
 pub use dense_polynomial::ExtendedDensePolynomial;
