+++ utils/src/math.rs
@@ -1,0 +1,15 @@
+/// Returns ceil(log2(d)) but panics if d = 0.
+
+pub fn ceil_log2(d: usize) -> usize {
+    assert!(d != 0);
+    let mut pow2 = 1;
+    let mut ceil_log2 = 0;
+    while d > pow2 {
+        ceil_log2 += 1;
+        pow2 = match pow2.checked_mul(2) {
+            Some(x) => x,
+            None => break,
+        }
+    }
+    ceil_log2
+}
